Make molecule reuse prerun from ansible-lint
As ansible-lint is able to detect repository layout and run prepare
ansible from running it, we rely on its logic in order to assure
ansible is able to run properly.

This means that dependencies will be installed into .cache folder
and that the role import path will be updated to include it.